requirements.txt
backcall==0.2.0
beautifulsoup4==4.6.0
blis==0.2.4
certifi==2021.5.30
charset-normalizer==2.0.4
ChatterBot==1.1.0
chatterbot-corpus==1.2.0
click==8.0.1
colorama==0.4.4
cymem==2.0.5
Cython==0.29.24
dataclasses==0.8
decorator==5.0.9
en-core-web-sm==2.1.0
entrypoints==0.3
Flask==2.0.1
idna==3.2
importlib-metadata==4.6.4
importlib-resources==5.2.2
ipykernel==5.5.5
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
JPype1==1.3.0
jupyter-client==7.0.1
jupyter-core==4.7.1
konlpy==0.5.2
lxml==4.6.3
MarkupSafe==2.0.1
mathparse==0.1.2
mecab-python===0.996-ko-0.9.2
murmurhash==1.0.5
nest-asyncio==1.5.1
nltk==3.6.2
numpy==1.19.5
oauthlib==3.1.1
packaging==21.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pint==0.17
pkg-resources==0.0.0
plac==0.9.6
preshed==2.0.1
prompt-toolkit==3.0.20
ptyprocess==0.7.0
Pygments==2.10.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.1
PyYAML==3.13
pyzmq==22.2.1
regex==2021.8.21
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
spacy==2.1.9
SQLAlchemy==1.3.24
srsly==1.0.5
thinc==7.0.8
tornado==6.1
tqdm==4.62.1
traitlets==4.3.3
tweepy==3.10.0
typing-extensions==3.10.0.0
urllib3==1.26.6
wasabi==0.8.2
wcwidth==0.2.5
Werkzeug==2.0.1
zipp==3.5.0
